backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
jinja2>=3.1.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.28.0
typing-extensions>=4.8.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.2
langchain-experimental>=0.0.7
beautifulsoup4>=4.12.0